requirements.txt
faiss-cpu==1.8.0
google-generativeai==0.4.1
huggingface-hub==0.22.2
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.36
langchain-google-genai==1.0.1
langchain-text-splitters==0.0.1
PyPDF2==3.0.1
streamlit==1.32.2
python-dotenv

